// appstate/types/merchantsTypes.ts
export enum MerchantsTypes {
	MERCHANTS_GET_REQUEST = "MERCHANTS_GET_REQUEST",
	MERCHANTS_GET_REQUEST_SUCCESS = "MERCHANTS_GET_REQUEST_SUCCESS",
	MERCHANTS_GET_REQUEST_UPDATE = "MERCHANTS_GET_REQUEST_UPDATE",
	MERCHANTS_GET_REQUEST_ERROR = "MERCHANTS_GET_REQUEST_ERROR",

	MERCHANTS_GET_SELECTED_MERCHANT = "MERCHANTS_GET_SELECTED_MERCHANT",

	MERCHANTS_DELETE_REQUEST = "MERCHANTS_DELETE_REQUEST",
	MERCHANTS_DELETE_REQUEST_SUCCESS = "MERCHANTS_DELETE_REQUEST_SUCCESS",
	MERCHANTS_DELETE_REQUEST_ERROR = "MERCHANTS_GET_REQUEST_ERROR"
}